+++ lib/edh/http_service/response.rb
@@ -1,9 +1,9 @@
-module Koala
+module EDH
   module HTTPService
     class Response
       attr_reader :status, :body, :headers
 
-      # Creates a new Response object, which standardizes the response received by Passport for use within Koala.
+      # Creates a new Response object, which standardizes the response received by Passport for use within EDH.
       def initialize(status, body, headers)
         @status = status
         @body = body
@@ -13,6 +13,6 @@ module Koala
   end
   
   # @private
-  # legacy support for when Response lived directly under Koala
+  # legacy support for when Response lived directly under EDH
   Response = HTTPService::Response
 end