+++ lib/edh/utils.rb
@@ -1,7 +1,7 @@
-module Koala
+module EDH
   module Utils
 
-    # Utility methods used by Koala.
+    # Utility methods used by EDH.
     require 'logger'
     require 'forwardable'
 
@@ -10,7 +10,7 @@ module Koala
 
     def_delegators :logger, :debug, :info, :warn, :error, :fatal, :level, :level=
 
-    # The Koala logger, an instance of the standard Ruby logger, pointing to STDOUT by default.
+    # The EDH logger, an instance of the standard Ruby logger, pointing to STDOUT by default.
     # In Rails projects, you can set this to Rails.logger.
     attr_accessor :logger
     self.logger = Logger.new(STDOUT)