+++ spec/cases/api_spec.rb
@@ -1,64 +1,64 @@
 require 'spec_helper'
 
-describe "Koala::Passport::API" do
+describe "EDH::Passport::API" do
   before(:each) do
-    @service = Koala::Passport::API.new
+    @service = EDH::Passport::API.new
   end
 
   it "doesn't include an access token if none was given" do
-    Koala.should_receive(:make_request).with(
+    EDH.should_receive(:make_request).with(
       anything,
       hash_not_including('access_token' => 1),
       anything,
       anything
-    ).and_return(Koala::HTTPService::Response.new(200, "", ""))
+    ).and_return(EDH::HTTPService::Response.new(200, "", ""))
 
     @service.api('anything')
   end
 
   it "includes an access token if given" do
     token = 'adfadf'
-    service = Koala::Passport::API.new token
+    service = EDH::Passport::API.new token
 
-    Koala.should_receive(:make_request).with(
+    EDH.should_receive(:make_request).with(
       anything,
       hash_including('access_token' => token),
       anything,
       anything
-    ).and_return(Koala::HTTPService::Response.new(200, "", ""))
+    ).and_return(EDH::HTTPService::Response.new(200, "", ""))
 
     service.api('anything')
   end
 
   it "has an attr_reader for access token" do
     token = 'adfadf'
-    service = Koala::Passport::API.new token
+    service = EDH::Passport::API.new token
     service.access_token.should == token
   end
 
-  it "gets the attribute of a Koala::HTTPService::Response given by the http_component parameter" do
+  it "gets the attribute of a EDH::HTTPService::Response given by the http_component parameter" do
     http_component = :method_name
 
-    response = mock('Mock KoalaResponse', :body => '', :status => 200)
+    response = mock('Mock EDHResponse', :body => '', :status => 200)
     result = stub("result")
     response.stub(http_component).and_return(result)
-    Koala.stub(:make_request).and_return(response)
+    EDH.stub(:make_request).and_return(response)
 
     @service.api('anything', {}, 'get', :http_component => http_component).should == result
   end
 
   it "returns the entire response if http_component => :response" do
     http_component = :response
-    response = mock('Mock KoalaResponse', :body => '', :status => 200)
-    Koala.stub(:make_request).and_return(response)
+    response = mock('Mock EDHResponse', :body => '', :status => 200)
+    EDH.stub(:make_request).and_return(response)
     @service.api('anything', {}, 'get', :http_component => http_component).should == response
   end
 
   it "turns arrays of non-enumerables into comma-separated arguments" do
     args = [12345, {:foo => [1, 2, "3", :four]}]
     expected = ["/12345", {:foo => "1,2,3,four"}, "get", {}]
-    response = mock('Mock KoalaResponse', :body => '', :status => 200)
-    Koala.should_receive(:make_request).with(*expected).and_return(response)
+    response = mock('Mock EDHResponse', :body => '', :status => 200)
+    EDH.should_receive(:make_request).with(*expected).and_return(response)
     @service.api(*args)
   end
 
@@ -69,14 +69,14 @@ describe "Koala::Passport::API" do
     # (if appropriate behavior is defined)
     # or raise an exception
     expected = ["/12345", params, "get", {}]
-    response = mock('Mock KoalaResponse', :body => '', :status => 200)
-    Koala.should_receive(:make_request).with(*expected).and_return(response)
+    response = mock('Mock EDHResponse', :body => '', :status => 200)
+    EDH.should_receive(:make_request).with(*expected).and_return(response)
     @service.api(*args)
   end
 
   it "returns the body of the request as JSON if no http_component is given" do
     response = stub('response', :body => 'body', :status => 200)
-    Koala.stub(:make_request).and_return(response)
+    EDH.stub(:make_request).and_return(response)
 
     json_body = mock('JSON body')
     MultiJson.stub(:load).and_return([json_body])
@@ -85,8 +85,8 @@ describe "Koala::Passport::API" do
   end
 
   it "executes an error checking block if provided" do
-    response = Koala::HTTPService::Response.new(200, '{}', {})
-    Koala.stub(:make_request).and_return(response)
+    response = EDH::HTTPService::Response.new(200, '{}', {})
+    EDH.stub(:make_request).and_return(response)
 
     yield_test = mock('Yield Tester')
     yield_test.should_receive(:pass)
@@ -98,46 +98,46 @@ describe "Koala::Passport::API" do
   end
 
   it "raises an API error if the HTTP response code is greater than or equal to 500" do
-    Koala.stub(:make_request).and_return(Koala::HTTPService::Response.new(500, 'response body', {}))
+    EDH.stub(:make_request).and_return(EDH::HTTPService::Response.new(500, 'response body', {}))
 
-    lambda { @service.api('anything') }.should raise_exception(Koala::Passport::APIError)
+    lambda { @service.api('anything') }.should raise_exception(EDH::Passport::APIError)
   end
 
   it "handles rogue true/false as responses" do
-    Koala.should_receive(:make_request).and_return(Koala::HTTPService::Response.new(200, 'true', {}))
+    EDH.should_receive(:make_request).and_return(EDH::HTTPService::Response.new(200, 'true', {}))
     @service.api('anything').should be_true
 
-    Koala.should_receive(:make_request).and_return(Koala::HTTPService::Response.new(200, 'false', {}))
+    EDH.should_receive(:make_request).and_return(EDH::HTTPService::Response.new(200, 'false', {}))
     @service.api('anything').should be_false
   end
 
   describe "with regard to leading slashes" do
     it "adds a leading / to the path if not present" do
       path = "anything"
-      Koala.should_receive(:make_request).with("/#{path}", anything, anything, anything).and_return(Koala::HTTPService::Response.new(200, 'true', {}))
+      EDH.should_receive(:make_request).with("/#{path}", anything, anything, anything).and_return(EDH::HTTPService::Response.new(200, 'true', {}))
       @service.api(path)
     end
 
     it "doesn't change the path if a leading / is present" do
       path = "/anything"
-      Koala.should_receive(:make_request).with(path, anything, anything, anything).and_return(Koala::HTTPService::Response.new(200, 'true', {}))
+      EDH.should_receive(:make_request).with(path, anything, anything, anything).and_return(EDH::HTTPService::Response.new(200, 'true', {}))
       @service.api(path)
     end
   end
 
   describe "with an access token" do
     before(:each) do
-      @api = Koala::Passport::API.new(@token)
+      @api = EDH::Passport::API.new(@token)
     end
 
-    it_should_behave_like "Koala RestAPI"
+    it_should_behave_like "EDH RestAPI"
   end
 
   describe "without an access token" do
     before(:each) do
-      @api = Koala::Passport::API.new
+      @api = EDH::Passport::API.new
     end
 
-    it_should_behave_like "Koala RestAPI"
+    it_should_behave_like "EDH RestAPI"
   end
 end