+++ spec/cases/edh_test_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
 
-describe KoalaTest do
+describe EDHTest do
   pending "should have tests, because the test suite depends on it"
 end