+++ spec/cases/error_spec.rb
@@ -1,24 +1,24 @@
 require 'spec_helper'
 
-describe Koala::Passport::APIError do
-  it "is a Koala::KoalaError" do
-    Koala::Passport::APIError.new(nil, nil).should be_a(Koala::KoalaError)
+describe EDH::Passport::APIError do
+  it "is a EDH::EDHError" do
+    EDH::Passport::APIError.new(nil, nil).should be_a(EDH::EDHError)
   end
 
   [:pp_error_type, :pp_error_code, :pp_error_subcode, :pp_error_message, :http_status, :response_body].each do |accessor|
     it "has an accessor for #{accessor}" do
-      Koala::Passport::APIError.instance_methods.map(&:to_sym).should include(accessor)
-      Koala::Passport::APIError.instance_methods.map(&:to_sym).should include(:"#{accessor}=")
+      EDH::Passport::APIError.instance_methods.map(&:to_sym).should include(accessor)
+      EDH::Passport::APIError.instance_methods.map(&:to_sym).should include(:"#{accessor}=")
     end
   end
 
   it "sets http_status to the provided status" do
     error_response = '{ "error": {"type": "foo", "other_details": "bar"} }'
-    Koala::Passport::APIError.new(400, error_response).response_body.should == error_response
+    EDH::Passport::APIError.new(400, error_response).response_body.should == error_response
   end
 
   it "sets response_body to the provided response body" do
-    Koala::Passport::APIError.new(400, '').http_status.should == 400
+    EDH::Passport::APIError.new(400, '').http_status.should == 400
   end
 
   context "with an error_info hash" do
@@ -29,7 +29,7 @@ describe Koala::Passport::APIError do
         'code' => 1,
         'error_subcode' => 'subcode'
       }
-      Koala::Passport::APIError.new(400, '', error_info)
+      EDH::Passport::APIError.new(400, '', error_info)
     }
 
     {
@@ -51,7 +51,7 @@ describe Koala::Passport::APIError do
   context "with an error_info string" do
     it "sets the error message \"error_info [HTTP http_status]\"" do
       error_info = "Passport is down."
-      error = Koala::Passport::APIError.new(400, '', error_info)
+      error = EDH::Passport::APIError.new(400, '', error_info)
       error.message.should == "Passport is down. [HTTP 400]"
     end
   end
@@ -59,7 +59,7 @@ describe Koala::Passport::APIError do
   context "with no error_info and a response_body containing error JSON" do
     it "should extract the error info from the response body" do
       response_body = '{ "error": { "type": "type", "message": "message", "code": 1, "error_subcode": "subcode" } }'
-      error = Koala::Passport::APIError.new(400, response_body)
+      error = EDH::Passport::APIError.new(400, response_body)
       {
         :pp_error_type => 'type',
         :pp_error_message => 'message',
@@ -73,32 +73,32 @@ describe Koala::Passport::APIError do
 
 end
 
-describe Koala::KoalaError do
+describe EDH::EDHError do
   it "is a StandardError" do
-     Koala::KoalaError.new.should be_a(StandardError)
+     EDH::EDHError.new.should be_a(StandardError)
   end
 end
 
-describe Koala::Passport::BadPassportResponse do
-  it "is a Koala::Passport::APIError" do
-     Koala::Passport::BadPassportResponse.new(nil, nil).should be_a(Koala::Passport::APIError)
+describe EDH::Passport::BadPassportResponse do
+  it "is a EDH::Passport::APIError" do
+     EDH::Passport::BadPassportResponse.new(nil, nil).should be_a(EDH::Passport::APIError)
   end
 end
 
-describe Koala::Passport::ServerError do
-  it "is a Koala::Passport::APIError" do
-     Koala::Passport::ServerError.new(nil, nil).should be_a(Koala::Passport::APIError)
+describe EDH::Passport::ServerError do
+  it "is a EDH::Passport::APIError" do
+     EDH::Passport::ServerError.new(nil, nil).should be_a(EDH::Passport::APIError)
   end
 end
 
-describe Koala::Passport::ClientError do
-  it "is a Koala::Passport::APIError" do
-     Koala::Passport::ClientError.new(nil, nil).should be_a(Koala::Passport::APIError)
+describe EDH::Passport::ClientError do
+  it "is a EDH::Passport::APIError" do
+     EDH::Passport::ClientError.new(nil, nil).should be_a(EDH::Passport::APIError)
   end
 end
 
-describe Koala::Passport::AuthenticationError do
-  it "is a Koala::Passport::ClientError" do
-     Koala::Passport::AuthenticationError.new(nil, nil).should be_a(Koala::Passport::ClientError)
+describe EDH::Passport::AuthenticationError do
+  it "is a EDH::Passport::ClientError" do
+     EDH::Passport::AuthenticationError.new(nil, nil).should be_a(EDH::Passport::ClientError)
   end
 end