+++ spec/cases/http_service_spec.rb
@@ -1,18 +1,18 @@
 require 'spec_helper'
 
-describe "Koala::HTTPService" do
+describe "EDH::HTTPService" do
   it "has a faraday_middleware accessor" do
-    Koala::HTTPService.methods.map(&:to_sym).should include(:faraday_middleware)
-    Koala::HTTPService.methods.map(&:to_sym).should include(:faraday_middleware=)
+    EDH::HTTPService.methods.map(&:to_sym).should include(:faraday_middleware)
+    EDH::HTTPService.methods.map(&:to_sym).should include(:faraday_middleware=)
   end
 
   it "has an http_options accessor" do
-    Koala::HTTPService.should respond_to(:http_options)
-    Koala::HTTPService.should respond_to(:http_options=)
+    EDH::HTTPService.should respond_to(:http_options)
+    EDH::HTTPService.should respond_to(:http_options=)
   end
 
   it "sets http_options to {} by default" do
-    Koala::HTTPService.http_options.should == {}
+    EDH::HTTPService.http_options.should == {}
   end
 
   describe "DEFAULT_MIDDLEWARE" do
@@ -24,30 +24,30 @@ describe "Koala::HTTPService" do
     end
 
     it "is defined" do
-      Koala::HTTPService.const_defined?("DEFAULT_MIDDLEWARE").should be_true
+      EDH::HTTPService.const_defined?("DEFAULT_MIDDLEWARE").should be_true
     end
 
     it "adds multipart" do
-      @builder.should_receive(:use).with(Koala::HTTPService::MultipartRequest)
-      Koala::HTTPService::DEFAULT_MIDDLEWARE.call(@builder)
+      @builder.should_receive(:use).with(EDH::HTTPService::MultipartRequest)
+      EDH::HTTPService::DEFAULT_MIDDLEWARE.call(@builder)
     end
 
     it "adds url_encoded" do
       @builder.should_receive(:request).with(:url_encoded)
-      Koala::HTTPService::DEFAULT_MIDDLEWARE.call(@builder)
+      EDH::HTTPService::DEFAULT_MIDDLEWARE.call(@builder)
     end
 
     it "uses the default adapter" do
       adapter = :testing_now
       Faraday.stub(:default_adapter).and_return(adapter)
       @builder.should_receive(:adapter).with(adapter)
-      Koala::HTTPService::DEFAULT_MIDDLEWARE.call(@builder)
+      EDH::HTTPService::DEFAULT_MIDDLEWARE.call(@builder)
     end
   end
 
   describe ".encode_params" do
     it "returns an empty string if param_hash evaluates to false" do
-      Koala::HTTPService.encode_params(nil).should == ''
+      EDH::HTTPService.encode_params(nil).should == ''
     end
 
     it "converts values to JSON if the value is not a String" do
@@ -63,7 +63,7 @@ describe "Koala::HTTPService" do
         string => string
       }
 
-      result = Koala::HTTPService.encode_params(args)
+      result = EDH::HTTPService.encode_params(args)
       result.split('&').find do |key_and_val|
         key_and_val.match("#{not_a_string}=#{val}")
       end.should be_true
@@ -72,7 +72,7 @@ describe "Koala::HTTPService" do
     it "escapes all values" do
       args = Hash[*(1..4).map {|i| [i.to_s, "Value #{i}($"]}.flatten]
 
-      result = Koala::HTTPService.encode_params(args)
+      result = EDH::HTTPService.encode_params(args)
       result.split('&').each do |key_val|
         key, val = key_val.split('=')
         val.should == CGI.escape(args[key])
@@ -82,14 +82,14 @@ describe "Koala::HTTPService" do
     it "encodes parameters in alphabetical order" do
       args = {:b => '2', 'a' => '1'}
 
-      result = Koala::HTTPService.encode_params(args)
+      result = EDH::HTTPService.encode_params(args)
       result.split('&').map{|key_val| key_val.split('=')[0]}.should == ['a', 'b']
     end
 
     it "converts all keys to Strings" do
       args = Hash[*(1..4).map {|i| [i, "val#{i}"]}.flatten]
 
-      result = Koala::HTTPService.encode_params(args)
+      result = EDH::HTTPService.encode_params(args)
       result.split('&').each do |key_val|
         key, val = key_val.split('=')
         key.should == args.find{|key_val_arr| key_val_arr.last == val}.first.to_s
@@ -113,77 +113,77 @@ describe "Koala::HTTPService" do
     describe "creating the Faraday connection" do
       it "creates a Faraday connection using the server" do
         server = "foo"
-        Koala::HTTPService.stub(:server).and_return(server)
+        EDH::HTTPService.stub(:server).and_return(server)
         Faraday.should_receive(:new).with(server, anything).and_return(@mock_connection)
-        Koala::HTTPService.make_request("anything", {}, "anything")
+        EDH::HTTPService.make_request("anything", {}, "anything")
       end
 
-      it "merges Koala::HTTPService.http_options into the request params" do
+      it "merges EDH::HTTPService.http_options into the request params" do
         http_options = {:a => 2, :c => "3"}
-        Koala::HTTPService.http_options = http_options
+        EDH::HTTPService.http_options = http_options
         Faraday.should_receive(:new).with(anything, hash_including(http_options)).and_return(@mock_connection)
-        Koala::HTTPService.make_request("anything", {}, "get")
+        EDH::HTTPService.make_request("anything", {}, "get")
       end
 
       it "merges any provided options into the request params" do
         options = {:a => 2, :c => "3"}
         Faraday.should_receive(:new).with(anything, hash_including(options)).and_return(@mock_connection)
-        Koala::HTTPService.make_request("anything", {}, "get", options)
+        EDH::HTTPService.make_request("anything", {}, "get", options)
       end
 
-      it "overrides Koala::HTTPService.http_options with any provided options for the request params" do
+      it "overrides EDH::HTTPService.http_options with any provided options for the request params" do
         options = {:a => 2, :c => "3"}
         http_options = {:a => :a}
-        Koala::HTTPService.stub(:http_options).and_return(http_options)
+        EDH::HTTPService.stub(:http_options).and_return(http_options)
 
         Faraday.should_receive(:new).with(anything, hash_including(http_options.merge(options))).and_return(@mock_connection)
-        Koala::HTTPService.make_request("anything", {}, "get", options)
+        EDH::HTTPService.make_request("anything", {}, "get", options)
       end
 
       it "allows you to set other verify modes if you really want" do
         options = {:ssl => {:verify => :foo}}
         Faraday.should_receive(:new).with(anything, hash_including(options)).and_return(@mock_connection)
-        Koala::HTTPService.make_request("anything", {"access_token" => "foo"}, "get", options)
+        EDH::HTTPService.make_request("anything", {"access_token" => "foo"}, "get", options)
       end
 
       it "calls server with the composite options" do
         options = {:a => 2, :c => "3"}
         http_options = {:a => :a}
-        Koala::HTTPService.stub(:http_options).and_return(http_options)
-        Koala::HTTPService.should_receive(:server).with(hash_including(http_options.merge(options))).and_return("foo")
-        Koala::HTTPService.make_request("anything", {}, "get", options)
+        EDH::HTTPService.stub(:http_options).and_return(http_options)
+        EDH::HTTPService.should_receive(:server).with(hash_including(http_options.merge(options))).and_return("foo")
+        EDH::HTTPService.make_request("anything", {}, "get", options)
       end
 
       it "uses the default builder block if HTTPService.faraday_middleware block is not defined" do
         block = Proc.new {}
-        stub_const("Koala::HTTPService::DEFAULT_MIDDLEWARE", block)
-        Koala::HTTPService.stub(:faraday_middleware).and_return(nil)
+        stub_const("EDH::HTTPService::DEFAULT_MIDDLEWARE", block)
+        EDH::HTTPService.stub(:faraday_middleware).and_return(nil)
         Faraday.should_receive(:new).with(anything, anything, &block).and_return(@mock_connection)
-        Koala::HTTPService.make_request("anything", {}, "get")
+        EDH::HTTPService.make_request("anything", {}, "get")
       end
 
       it "uses the defined HTTPService.faraday_middleware block if defined" do
         block = Proc.new { }
-        Koala::HTTPService.should_receive(:faraday_middleware).and_return(block)
+        EDH::HTTPService.should_receive(:faraday_middleware).and_return(block)
         Faraday.should_receive(:new).with(anything, anything, &block).and_return(@mock_connection)
-        Koala::HTTPService.make_request("anything", {}, "get")
+        EDH::HTTPService.make_request("anything", {}, "get")
       end
     end
 
     it "makes a POST request if the verb isn't get" do
       @mock_connection.should_receive(:post).and_return(@mock_http_response)
-      Koala::HTTPService.make_request("anything", {}, "anything")
+      EDH::HTTPService.make_request("anything", {}, "anything")
     end
 
     it "includes the verb in the body if the verb isn't get" do
       verb = "eat"
       @mock_connection.should_receive(:post).with(anything, hash_including("method" => verb)).and_return(@mock_http_response)
-      Koala::HTTPService.make_request("anything", {}, verb)
+      EDH::HTTPService.make_request("anything", {}, verb)
     end
 
     it "makes a GET request if the verb is get" do
       @mock_connection.should_receive(:get).and_return(@mock_http_response)
-      Koala::HTTPService.make_request("anything", {}, "get")
+      EDH::HTTPService.make_request("anything", {}, "get")
     end
 
     describe "for GETs" do
@@ -191,20 +191,20 @@ describe "Koala::HTTPService" do
         # technically this is done for all requests, but you don't send GET requests with files
         args = {"a" => :b, "c" => 3}
         Faraday.should_receive(:new).with(anything, hash_including(:params => args)).and_return(@mock_connection)
-        Koala::HTTPService.make_request("anything", args, "get")
+        EDH::HTTPService.make_request("anything", args, "get")
       end
 
       it "submits nothing to the body" do
         # technically this is done for all requests, but you don't send GET requests with files
         args = {"a" => :b, "c" => 3}
         @mock_connection.should_receive(:get).with(anything, {}).and_return(@mock_http_response)
-        Koala::HTTPService.make_request("anything", args, "get")
+        EDH::HTTPService.make_request("anything", args, "get")
       end
 
       it "logs verb, url and params to debug" do
         args = {"a" => :b, "c" => 3}
         log_message_stem = "GET: anything params: "
-        Koala::Utils.logger.should_receive(:debug) do |log_message|
+        EDH::Utils.logger.should_receive(:debug) do |log_message|
           # unordered hashes are a bane
           # Ruby in 1.8 modes tends to return different hash orderings,
           # which makes checking the content of the stringified hash hard
@@ -213,7 +213,7 @@ describe "Koala::HTTPService" do
           log_message.match(/\{.*\}/).should_not be_nil
         end
 
-        Koala::HTTPService.make_request("anything", args, "get")
+        EDH::HTTPService.make_request("anything", args, "get")
       end
     end
 
@@ -222,13 +222,13 @@ describe "Koala::HTTPService" do
         # technically this is done for all requests, but you don't send GET requests with files
         args = {"a" => :b, "c" => 3}
         @mock_connection.should_receive(:post).with(anything, hash_including(args)).and_return(@mock_http_response)
-        Koala::HTTPService.make_request("anything", args, "post")
+        EDH::HTTPService.make_request("anything", args, "post")
       end
 
       it "logs verb, url and params to debug" do
         args = {"a" => :b, "c" => 3}
         log_message_stem = "POST: anything params: "
-        Koala::Utils.logger.should_receive(:debug) do |log_message|
+        EDH::Utils.logger.should_receive(:debug) do |log_message|
           # unordered hashes are a bane
           # Ruby in 1.8 modes tends to return different hash orderings,
           # which makes checking the content of the stringified hash hard
@@ -236,19 +236,19 @@ describe "Koala::HTTPService" do
           log_message.should include(log_message_stem)
           log_message.match(/\{.*\}/).should_not be_nil
         end
-        Koala::HTTPService.make_request("anything", args, "post")
+        EDH::HTTPService.make_request("anything", args, "post")
       end
     end
   end
 
   describe "deprecated options" do
     before :each do
-      Koala::HTTPService.stub(:http_options).and_return({})
-      @service = Koala.http_service
+      EDH::HTTPService.stub(:http_options).and_return({})
+      @service = EDH.http_service
     end
 
     after :each do
-      Koala.http_service = @service
+      EDH.http_service = @service
     end
 
     describe "per-request options" do
@@ -268,13 +268,13 @@ describe "Koala::HTTPService" do
         it "merges any typhoeus_options into options" do
           typhoeus_options = {:a => 2}
           Faraday.should_receive(:new).with(anything, hash_including(typhoeus_options)).and_return(@mock_connection)
-          Koala::HTTPService.make_request("anything", {}, "get", :typhoeus_options => typhoeus_options)
+          EDH::HTTPService.make_request("anything", {}, "get", :typhoeus_options => typhoeus_options)
         end
 
         it "deletes the typhoeus_options key" do
           typhoeus_options = {:a => 2}
           Faraday.should_receive(:new).with(anything, hash_not_including(:typhoeus_options => typhoeus_options)).and_return(@mock_connection)
-          Koala::HTTPService.make_request("anything", {}, "get", :typhoeus_options => typhoeus_options)
+          EDH::HTTPService.make_request("anything", {}, "get", :typhoeus_options => typhoeus_options)
         end
       end
 
@@ -282,13 +282,13 @@ describe "Koala::HTTPService" do
         it "sets any ca_path into options[:ssl]" do
           ca_path = :foo
           Faraday.should_receive(:new).with(anything, hash_including(:ssl => hash_including(:ca_path => ca_path))).and_return(@mock_connection)
-          Koala::HTTPService.make_request("anything", {}, "get", :ca_path => ca_path)
+          EDH::HTTPService.make_request("anything", {}, "get", :ca_path => ca_path)
         end
 
         it "deletes the ca_path key" do
           ca_path = :foo
           Faraday.should_receive(:new).with(anything, hash_not_including(:ca_path => ca_path)).and_return(@mock_connection)
-          Koala::HTTPService.make_request("anything", {}, "get", :ca_path => ca_path)
+          EDH::HTTPService.make_request("anything", {}, "get", :ca_path => ca_path)
         end
       end
 
@@ -296,13 +296,13 @@ describe "Koala::HTTPService" do
         it "sets any ca_file into options[:ssl]" do
           ca_file = :foo
           Faraday.should_receive(:new).with(anything, hash_including(:ssl => hash_including(:ca_file => ca_file))).and_return(@mock_connection)
-          Koala::HTTPService.make_request("anything", {}, "get", :ca_file => ca_file)
+          EDH::HTTPService.make_request("anything", {}, "get", :ca_file => ca_file)
         end
 
         it "deletes the ca_file key" do
           ca_file = :foo
           Faraday.should_receive(:new).with(anything, hash_not_including(:ca_file => ca_file)).and_return(@mock_connection)
-          Koala::HTTPService.make_request("anything", {}, "get", :ca_file => ca_file)
+          EDH::HTTPService.make_request("anything", {}, "get", :ca_file => ca_file)
         end
       end
 
@@ -310,13 +310,13 @@ describe "Koala::HTTPService" do
         it "sets any verify_mode into options[:ssl]" do
           verify_mode = :foo
           Faraday.should_receive(:new).with(anything, hash_including(:ssl => hash_including(:verify_mode => verify_mode))).and_return(@mock_connection)
-          Koala::HTTPService.make_request("anything", {}, "get", :verify_mode => verify_mode)
+          EDH::HTTPService.make_request("anything", {}, "get", :verify_mode => verify_mode)
         end
 
         it "deletes the verify_mode key" do
           verify_mode = :foo
           Faraday.should_receive(:new).with(anything, hash_not_including(:verify_mode => verify_mode)).and_return(@mock_connection)
-          Koala::HTTPService.make_request("anything", {}, "get", :verify_mode => verify_mode)
+          EDH::HTTPService.make_request("anything", {}, "get", :verify_mode => verify_mode)
         end
       end
     end