+++ changelog.md
@@ -1,4 +1,4 @@
 v0.1
 ====
 
-* Initial rework of Koala gem to work for Passport API, based off Koala gem by Alex Koppel.
+* Initial rework of EDH gem to work for Passport API, based off Koala gem by Alex Koppel.